Fix: Correct typo in the name of the binary in service block
When the homebrew formula was updated to auto-generate the `service`
block to support `brew services` (see
`8faa5a029f3ca07eb651e6058bb6374d8dfc680d`), there was a small typo in
the binary.

The binary should be `openssl-osx-ca`, not `openssl-ox-ca`. As a result,
whenever anyone tried to run `brew services start openssl-osx-ca`, it
would silently fail.

This corrects that.