Fix: Explicitly pass path to cert.pem in service definition
It seems that when we switched from a plist definition to a service
defintion block to be in line with homebrew's recommendations, we never
explicitly provided the path to the `cert.pem` file generated from the
keychain to the underlying binary `osx-ca-certs`.

The result was that an empty `cert.pem` was generated and used for
openssl. This....predictably broke configs.

All credit goes to @akelge for this fix as she did all the legwork :)

This should close: https://github.com/raggi/openssl-osx-ca/issues/27